test: split long running test
